+++ mistralrs-quant/src/cublaslt/matmul.rs
@@ -413,27 +413,25 @@ pub trait Matmul<T: CublasLTDType>: MatmulShared {
         let matmul_desc = MatmulDesc::new(
             sys::cublasComputeType_t::CUBLAS_COMPUTE_32F,
             sys::cudaDataType_t::CUDA_R_32F,
-        )
-        .unwrap();
+        )?;
 
         // Set transa
-        matmul_desc.set_transpose(cfg.transa, Matrix::A).unwrap();
+        matmul_desc.set_transpose(cfg.transa, Matrix::A)?;
         // Set transb
-        matmul_desc.set_transpose(cfg.transb, Matrix::B).unwrap();
+        matmul_desc.set_transpose(cfg.transb, Matrix::B)?;
 
         // Creates matrix layouts
-        let a_layout = MatrixLayout::new(Self::matrix_type(), a_rows, a_cols, cfg.lda).unwrap();
+        let a_layout = MatrixLayout::new(Self::matrix_type(), a_rows, a_cols, cfg.lda)?;
         if let (Some(batch_size), Some(stride_a)) = (cfg.batch_size, cfg.stride_a) {
             a_layout.set_batch(batch_size, stride_a)?;
         }
 
-        let b_layout = MatrixLayout::new(Self::matrix_type(), b_rows, b_cols, cfg.ldb).unwrap();
+        let b_layout = MatrixLayout::new(Self::matrix_type(), b_rows, b_cols, cfg.ldb)?;
         if let (Some(batch_size), Some(stride_b)) = (cfg.batch_size, cfg.stride_b) {
             b_layout.set_batch(batch_size, stride_b)?;
         }
 
-        let c_layout =
-            MatrixLayout::new(sys::cudaDataType_t::CUDA_R_16BF, cfg.m, cfg.n, cfg.ldc).unwrap();
+        let c_layout = MatrixLayout::new(sys::cudaDataType_t::CUDA_R_16BF, cfg.m, cfg.n, cfg.ldc)?;
         if let (Some(batch_size), Some(stride_c)) = (cfg.batch_size, cfg.stride_c) {
             c_layout.set_batch(batch_size, stride_c)?;
         }
@@ -442,21 +440,15 @@ pub trait Matmul<T: CublasLTDType>: MatmulShared {
             OutSlice::F8(_) => Self::matrix_type(),
             OutSlice::BF16(_) => sys::cudaDataType_t::CUDA_R_16BF,
         };
-        let d_layout = MatrixLayout::new(out_ty, cfg.m, cfg.n, cfg.ldc).unwrap();
+        let d_layout = MatrixLayout::new(out_ty, cfg.m, cfg.n, cfg.ldc)?;
         if let (Some(batch_size), Some(stride_c)) = (cfg.batch_size, cfg.stride_c) {
             d_layout.set_batch(batch_size, stride_c)?;
         }
 
         // Set scale factors
-        matmul_desc
-            .set_scale_ptr(scale_a.device_ptr(), Matrix::A)
-            .unwrap();
-        matmul_desc
-            .set_scale_ptr(scale_b.device_ptr(), Matrix::B)
-            .unwrap();
-        matmul_desc
-            .set_scale_ptr(scale_d.device_ptr(), Matrix::D)
-            .unwrap();
+        matmul_desc.set_scale_ptr(scale_a.device_ptr(), Matrix::A)?;
+        matmul_desc.set_scale_ptr(scale_b.device_ptr(), Matrix::B)?;
+        matmul_desc.set_scale_ptr(scale_d.device_ptr(), Matrix::D)?;
 
         // Pass amaxd ptr
         // unsafe {
@@ -466,21 +458,17 @@ pub trait Matmul<T: CublasLTDType>: MatmulShared {
         //         amax_d.device_ptr_mut() as *const CUdeviceptr as *const _,
         //         mem::size_of::<CUdeviceptr>(),
         //     )
-        //     .unwrap();
+        //     ?;
         // }
 
         // Epilogue system can be leveraged to fuse add and activation operations
-        matmul_desc
-            .set_epilogue(act, bias.map(|b| b.device_ptr()), cfg.stride_bias)
-            .unwrap();
+        matmul_desc.set_epilogue(act, bias.map(|b| b.device_ptr()), cfg.stride_bias)?;
 
         // Create matmul heuristic search preferences
-        let matmul_pref = MatmulPref::new().unwrap();
+        let matmul_pref = MatmulPref::new()?;
 
         // Set workspace size
-        matmul_pref
-            .set_workspace_size(self.workspace().size)
-            .unwrap();
+        matmul_pref.set_workspace_size(self.workspace().size)?;
 
         // Get heuristic given Config, bias, act and workspace size
         let heuristic = result::get_matmul_algo_heuristic(
@@ -491,8 +479,7 @@ pub trait Matmul<T: CublasLTDType>: MatmulShared {
             c_layout.handle,
             d_layout.handle,
             matmul_pref.handle,
-        )
-        .unwrap();
+        )?;
 
         let out_ptr = match out {
             OutSlice::BF16(s) => s.device_ptr_mut(),
