+++ mistralrs-quant/src/cublaslt/mod.rs
@@ -11,6 +11,8 @@ use std::sync::{Mutex, Once};
 mod api;
 #[cfg(feature = "cuda")]
 mod matmul;
+#[cfg(test)]
+mod tests;
 
 #[cfg(feature = "cuda")]
 pub use api::{fused_batch_matmul, fused_batch_matmul_f8, CublasLt};
